Move business logic to interactor
